gz-backend/requirement.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.4.0
Flask-Cors==4.0.1
